fix(qoder): keep credit used as a full float
int() truncated chat credits (2.887… → 2) in QuotaItem and cache.
Store the exact float; tracker UI may show two decimal places.

+++ backend/app/routers/quota.py
@@ -82,8 +82,8 @@ class QuotaItem(BaseModel):
     """A single quota metric for a provider connection."""
 
     name: str
-    used: int
-    total: int
+    used: float
+    total: float
     reset_at: Optional[str] = None
     remaining_percentage: float
 
@@ -150,8 +150,8 @@ def _parse_cached_quotas(raw: str | None) -> list[QuotaItem]:
             items.append(
                 QuotaItem(
                     name=str(q.get("name") or ""),
-                    used=int(q.get("used") or 0),
-                    total=int(q.get("total") or 0),
+                    used=float(q.get("used") or 0),
+                    total=float(q.get("total") or 0),
                     reset_at=q.get("reset_at"),
                     remaining_percentage=float(
                         q.get("remaining_percentage") or 0
